[development]...

+++ src/br/ufrpe/gerenciadorderelatorios/model/BancoDeDadosGeRel.java
@@ -8,6 +8,7 @@ import java.io.Serializable;
 
 import br.ufrpe.gerenciadorderelatorios.excecoes.ArquivoOuDiretorioNaoExisteException;
 import br.ufrpe.gerenciadorderelatorios.excecoes.ExclusaoDeArquivoOuDiretorioNegadaException;
+import br.ufrpe.gerenciadorderelatorios.excecoes.JaExisteArquivoOuDiretorioException;
 import br.ufrpe.gerenciadorderelatorios.excecoes.TipoDeArquivoDifereDoEsperadoException;
 
 public class BancoDeDadosGeRel {
@@ -49,45 +50,58 @@ public class BancoDeDadosGeRel {
 		}
 	}
 	
-	/**Salva o objeto no diretório de acordo com o seu tipo.*/
-	public void adicionar(File base, Estrutura estrutura , Gravavel gravavel) {
-		String caminho =  this.construirCaminho(new String[] {base.getAbsolutePath(), estrutura.obterRaiz()});
-		File diretorio = new File(caminho);
+	/**Salva o objeto gravavel no diretório apontado pela estrutura.
+	 * @throws JaExisteArquivoOuDiretorioException */
+	public void adicionar(File base, Estrutura estrutura , Gravavel gravavel) throws JaExisteArquivoOuDiretorioException {
+		String caminhoDiretorio =  this.construirCaminho(new String[] {base.getAbsolutePath(), estrutura.obterRaiz()});
+		File diretorio = new File(caminhoDiretorio);
+		
 		/*Verificando se o hisórico já existe.*/
 		if (diretorio.exists() && diretorio.isDirectory()) {
+			
 			/*Verificando se já esta no diretório correto para salvar.*/
 			if(estrutura.obterSubDiretorios() == null) {
 				String caminhoArquivo = this.construirCaminho(new String[] {diretorio.getAbsolutePath(), gravavel.obterId()+".ser"});
-				FileOutputStream arquivoSaida = null;
-				ObjectOutputStream objetoSaida = null;
+				File file = new File(caminhoArquivo);
 				
-				try {
-					arquivoSaida = new FileOutputStream(caminhoArquivo);
-					objetoSaida = new ObjectOutputStream(arquivoSaida);
-					objetoSaida.writeObject(gravavel);
-				} catch (IOException e) {
-					e.printStackTrace();
-				} finally {
-					if (arquivoSaida != null) {
-						try {
-							arquivoSaida.close();
-						} catch (IOException e) {
-							e.printStackTrace();
+				/*Verificando se o arquivo ainda não existe.*/
+				if(!file.exists()) {
+					FileOutputStream arquivoSaida = null;
+					ObjectOutputStream objetoSaida = null;
+					
+					try {
+						arquivoSaida = new FileOutputStream(caminhoArquivo);
+						objetoSaida = new ObjectOutputStream(arquivoSaida);
+						objetoSaida.writeObject(gravavel);
+					} catch (IOException e) {
+						e.printStackTrace();
+					} finally {
+						if (arquivoSaida != null) {
+							try {
+								arquivoSaida.close();
+							} catch (IOException e) {
+								e.printStackTrace();
+							}
+						}
+						if (objetoSaida != null) {
+							try {
+								objetoSaida.close();
+							} catch (IOException e) {
+								e.printStackTrace();
+							}
 						}
 					}
-					if (objetoSaida != null) {
-						try {
-							objetoSaida.close();
-						} catch (IOException e) {
-							e.printStackTrace();
-						}
-					}
+					
+				} else {
+					throw new JaExisteArquivoOuDiretorioException("O arquivo já existe no banco e dados em "+estrutura.obterRaiz());
 				}
+				
 			} else {
 				for(Estrutura e: estrutura.obterSubDiretorios()) {
-					this.adicionar(base, e, gravavel);
+					this.adicionar(diretorio, e, gravavel);
 				}
 			}
+			
 		} else {
 			criarEstrutura(base, estrutura);
 			this.adicionar(base, estrutura, gravavel);
@@ -97,23 +111,36 @@ public class BancoDeDadosGeRel {
 	/**Remove o arquivo ou diretório especificado se houver.
 	 * @throws ArquivoOuDiretorioNaoExisteException 
 	 * @throws ExclusaoDeArquivoOuDiretorioNegadaException*/
-	public void remover(String diretorioPai, String nomeArquivo) throws ArquivoOuDiretorioNaoExisteException, ExclusaoDeArquivoOuDiretorioNegadaException {
-		File diretorio = this.buscarDiretorioArquivo(this.obterBancoDeDados(), diretorioPai);
-		/*Verificando se diretório existe*/
-		if (diretorio != null) {
-			String caminhoArquivo = this.construirCaminho(new String[] {diretorio.getAbsolutePath(), nomeArquivo});
-			/*Verificando se a estrutura esta completa para adicionar o arquivo*/
-			File alvo = new File(caminhoArquivo);
-			/*Verificando se já existe o arquivo*/
-			if(alvo.exists()) {
-				if(!alvo.delete()) {
-					throw new ExclusaoDeArquivoOuDiretorioNegadaException("Permissao para excluir o arquivo foi negada ou diretório não esta vazio: "+alvo.obterName());
+	public void remover(File base, Estrutura estrutura , String nomeArquivo) throws ArquivoOuDiretorioNaoExisteException, ExclusaoDeArquivoOuDiretorioNegadaException {
+		String caminhoDiretorio =  this.construirCaminho(new String[] {base.getAbsolutePath(), estrutura.obterRaiz()});
+		File diretorio = new File(caminhoDiretorio);
+		
+		/*Verificando se o diretório existe*/
+		if (diretorio.exists()) {
+			
+			/*Verificando se já é esta o diretório correto para exlusão.*/
+			if(estrutura.obterSubDiretorios() == null) {
+				String caminhoArquivo = this.construirCaminho(new String[] {diretorio.getAbsolutePath(), nomeArquivo});
+				File alvo = new File(caminhoArquivo);
+				
+				/*Verificando se o arquivo existe.*/
+				if(alvo.exists()) {
+					
+					/*Deletando o arquivo ou diretório.*/
+					if(!alvo.delete()) {
+						throw new ExclusaoDeArquivoOuDiretorioNegadaException("Permissao para excluir o arquivo ou diretório negada: "+alvo.getName());
+					}
+					
+				} else {
+					throw new ArquivoOuDiretorioNaoExisteException("O diretório ou arquivo não existe: "+alvo.getName());
 				}
 			} else {
-				throw new ArquivoOuDiretorioNaoExisteException("O diretório ou arquivo não existe: "+alvo.getName());
+				for(Estrutura e: estrutura.obterSubDiretorios()) {
+					this.remover(diretorio, e, nomeArquivo);
+				}
 			}
 		} else {
-			throw new ArquivoOuDiretorioNaoExisteException("O diretorio não existe: "+diretorioPai);
+			throw new ArquivoOuDiretorioNaoExisteException("O diretorio não existe: "+estrutura.obterRaiz());
 		}
 	}
 	
@@ -133,7 +160,7 @@ public class BancoDeDadosGeRel {
 				if(arquivo.isFile()) {
 					alvo = arquivo;
 				} else {
-					throw new TipoDeArquivoDifereDoEsperadoException("Permissao para excluir o arquivo foi negada ou diretório não esta vazio: "+arquivo.obterName());
+					throw new TipoDeArquivoDifereDoEsperadoException("Permissao para excluir o arquivo foi negada ou diretório não esta vazio: "+arquivo.getName());
 				}
 			} else {
 				throw new ArquivoOuDiretorioNaoExisteException("O diretório ou arquivo não existe: "+arquivo.getName());
@@ -156,7 +183,7 @@ public class BancoDeDadosGeRel {
 	
 	@Override
 	public String toString() {
-		return "BancoDeDadosGeRel [diretorioDeTrabalho=" + diretorioDeTrabalho + ", bancoDeDados=" + bancoDeDados.getAbsolutePath() + "]";
+		return "BancoDeDadosGeRel [bancoDeDados=" + bancoDeDados + ", estrutura=" + estrutura + "]";
 	}
 
 	public File obterBancoDeDados() {
@@ -164,11 +191,11 @@ public class BancoDeDadosGeRel {
 	}
 
 	public Estrutura obterEstrutura() {
-		return bancos;
+		return this.estrutura;
 	}
 
-	public void definirEstrutura(Estrutura bancos) {
-		this.bancos = bancos;
+	public void definirEstrutura(Estrutura estrutura) {
+		this.estrutura = estrutura;
 	}
 
 }
